@@ -24,18 +24,16 @@ package zsawyer.mods.mumblelink;
 
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.api.distmarker.OnlyIn;
+import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.eventbus.api.EventPriority;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
-import net.minecraftforge.fml.ExtensionPoint;
+import net.minecraftforge.fml.IExtensionPoint;
 import net.minecraftforge.fml.ModLoadingContext;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.config.ModConfig;
 import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
-import net.minecraftforge.fml.network.FMLNetworkConstants;
 import net.minecraftforge.forgespi.language.IModInfo;
-
-import org.apache.commons.lang3.tuple.Pair;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import zsawyer.mods.mumblelink.api.MumbleLink;
@@ -61,7 +59,7 @@ public class MumbleLinkImpl extends MumbleLinkBase implements MumbleLink {
     private MumbleLinkAPIImpl api;
 
     public MumbleLinkImpl() {
-        FMLJavaModLoadingContext.get().getModEventBus().addListener(this::setup);
+        FMLJavaModLoadingContext.get().getModEventBus().register(this);
         this.preInit();
         instance = this;
     }
@@ -85,8 +83,10 @@ public class MumbleLinkImpl extends MumbleLinkBase implements MumbleLink {
 
     public void preInit() {
         ModLoadingContext context = ModLoadingContext.get();
-        context.registerExtensionPoint(ExtensionPoint.DISPLAYTEST, () -> Pair.of(
-            () -> FMLNetworkConstants.IGNORESERVERONLY, (serverVer, isDedicated) -> true));
+        context.registerExtensionPoint(IExtensionPoint.DisplayTest.class
+                , () -> new IExtensionPoint.DisplayTest(
+                        () -> "ANY",
+                        (serverVer, isDedicated) -> true));
         IModInfo modInfo = context.getActiveContainer().getModInfo();
         name = modInfo.getDisplayName();
         version = modInfo.getVersion().getQualifier();
