@@ -25,6 +25,7 @@ package zsawyer.mods.mumblelink.addons.pa.es;
 import net.minecraft.client.Minecraft;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.api.distmarker.OnlyIn;
+import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.eventbus.api.EventPriority;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.*;
@@ -32,9 +33,7 @@ import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.config.ModConfig;
 import net.minecraftforge.fml.event.lifecycle.InterModEnqueueEvent;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
-import net.minecraftforge.fml.network.FMLNetworkConstants;
 import net.minecraftforge.forgespi.language.IModInfo;
-import org.apache.commons.lang3.tuple.Pair;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import zsawyer.mods.mumblelink.api.Activateable;
@@ -80,7 +79,7 @@ public class ExtendedPASupport implements Activateable, IdentityManipulator {
     private MumbleLink mumbleLinkInstance;
 
     public ExtendedPASupport() {
-        FMLJavaModLoadingContext.get().getModEventBus().addListener(EventPriority.LOWEST, this::setup);
+        FMLJavaModLoadingContext.get().getModEventBus().register(this);
         this.preInit();
     }
 
@@ -103,8 +102,10 @@ public class ExtendedPASupport implements Activateable, IdentityManipulator {
 
     public void preInit() {
         ModLoadingContext context = ModLoadingContext.get();
-        context.registerExtensionPoint(ExtensionPoint.DISPLAYTEST, () -> Pair.of(
-            () -> FMLNetworkConstants.IGNORESERVERONLY, (serverVer, isDedicated) -> true));
+        context.registerExtensionPoint(IExtensionPoint.DisplayTest.class
+                , () -> new IExtensionPoint.DisplayTest(
+                        () -> "ANY",
+                        (serverVer, isDedicated) -> true));
         IModInfo modInfo = ModLoadingContext.get().getActiveContainer().getModInfo();
         name = modInfo.getDisplayName();
         version = modInfo.getVersion().getQualifier();
@@ -170,14 +171,14 @@ public class ExtendedPASupport implements Activateable, IdentityManipulator {
         // it connects to.
         // TODO: test if we can use game.world.getSeed()
         JSONArray spawnCoordinates = new JSONArray();
-        spawnCoordinates.put(game.world.getWorldInfo().getSpawnX());
-        spawnCoordinates.put(game.world.getWorldInfo().getSpawnY());
-        spawnCoordinates.put(game.world.getWorldInfo().getSpawnZ());
+        spawnCoordinates.put(game.level.getSharedSpawnPos().getX());
+        spawnCoordinates.put(game.level.getSharedSpawnPos().getX());
+        spawnCoordinates.put(game.level.getSharedSpawnPos().getX());
         // append coordinates
         identity.put(IdentityKey.WORLD_SPAWN, spawnCoordinates);
 
         // append the dimension
-        identity.put(IdentityKey.DIMENSION, game.player.getEntityWorld().getDimensionKey());
+        identity.put(IdentityKey.DIMENSION, game.player.level.dimension());
     }
 
     /**
